# internal/adapters/api/tag/response.go
package tag

import (
	"github.com/google/uuid"
	"time"
)

type TagResponse struct {
	ID        uuid.UUID `json:"id"`
	Title     string    `json:"title"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}